+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # Changelog
 
+## [ 2024/02/29 - v1.1.0 ]
+
+### BREAKING CHANGES
+
+* changed `cloud` prefix to `infra` in flags and config.
+* changed `build.nvidia` prefix to `build.gpu` in flags and config.
+* changed `build.nvidia.enable-nvidia-support` prefix to `build.gpu.enable-gpu-support` in flags and config.
+
+### Changed/Added
+
+* Added KubeVirt as a build option.
+* Supports AMD GPUs
+
 ## [ 2024/02/15 - v1.0.0 ]
 
 ### Changed/Added
+
 First release with:
 
 * Functioning support for OpenStack build, scan and signing.
-* Baski Server
+* Baski Server